requirements.txt
async-timeout==4.0.2
certifi==2022.9.14
charset-normalizer==2.1.1
Deprecated==1.2.13
edgedb==0.24.0
fire==0.4.0
idna==3.4
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.20.0
termcolor==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
wrapt==1.14.1
